pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "oqla-alrefai",email = "realrefai.rr@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "django-cloudinary-storage (>=0.3.0,<0.4.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
